R163 lint cleanup batch 2 (50+ files, 475 actionable warnings -> 0)
R162 cleaned 7 crates / 585 warnings. R163 continues with the remaining
crates. After this batch, only 30 `use of deprecated trait mock_llm::MockLlmProvider`
warnings remain - intentional deprecations signaling migration to
`apeireth_api::llm::LlmProvider + LlmAdvisorBackend`.

## Crates cleaned

| crate                          | before | after |
| ------------------------------ | -----: | ----: |
| apeireth-tool-fetch            |    157 |     0 |
| apeireth-memory                |    232 |     0 |
| apeireth-state                 |     26 |     0 |
| apeireth-council               |     24 |     0 |
| apeireth-sovereignty           |     15 |     0 |
| apeireth-provider              |     16 |     0 |
| apeireth-naming-v05            |     12 |     0 |
| apeireth-mcp                   |      6 |     0 |
| apeireth-tui                   |      3 |     1 |
| apeireth-value                 |      2 |     0 |
| apeireth-supervisor            |      1 |     0 |
| TOTAL actionable               |    475 |     0 |

## Bugs fixed (16 total)

- 12 unused vars prefixed with `_` (bilibili.rs url, identity.rs exists,
  user_profile.rs o, dailynote/enhanced.rs t, onion_consistency.rs cmp,
  synthesis.rs weights, ha.rs p/human_id/now_ms, three_domain_enforce.rs
  now_ms, backend.rs:1682 store in chat() fn, settings.rs enabled)
- 3 trivial casts removed (identity.rs:66 birth_time as i64, heartbeat.rs:354
  double u64 cast, main.rs:1103 u16 as u16)
- 1 unreachable statement restructured (onion_consistency.rs:193 - the
  `let _ = cmp;` after each match arm returns was unreachable; restructured
  to `match ... ;` where each arm returns/continues)
- MCP non_snake_case: `#![allow(non_snake_case)]` on lib.rs/tool_bridge.rs/
  initialize.rs because MCP/JSON-RPC wire protocol REQUIRES camelCase fields
  (protocolVersion / serverInfo / listChanged / inputSchema / clientInfo)

## Tests (cumulative +0 new, all pre-existing pass)

- apeireth-sovereignty/evidence_guard: 8 passed
- apeireth-sovereignty/three_domain_enforce: 14 passed
- apeireth-council/advisors::ethics: 5 passed
- apeireth-council/mock_llm: 3 passed
- apeireth-memory/identity: 11 passed
- apeireth-memory/user_profile: 9 passed
- apeireth-memory/dailynote::enhanced: 3 passed
- apeireth-value/onion_consistency: 12 passed
- apeireth-tool-fetch/bilibili: 6 passed
- apeireth-tui/pages::settings: 0 (no tests in module)

## Notes for future

- apeireth-supervisor::heartbeat::tests::t06_periodic_tick is a pre-existing
  timing-flaky test (asserts 2-5 ticks in 180ms; observed 1 tick). Per master
  directive, not addressed in this batch. See docs/r163.

## Verified

- cargo check --workspace: 0 errors
- R163 cleaned crates: 0 warnings each
- 30 remaining warnings are all intentional `MockLlmProvider` deprecation

## 0-touch statement

- 0 touches workspace.version (1.2.0)
- 0 touches 3 immutable spines (Self-Disable / L0 HA / 13-key verdict cache)
- 0 changes to docs/v4 / v4.1 / v2 / V0.5 / V1136 / 9键原始 / 18份stage2 / 14份stage3
- 0 引外部 dep
- 0 contains competitor name in public type/fn/mod/Cargo.toml description names

## Refs

- docs/r163/r163-lint-cleanup-batch-2.md
- 11 crate READMEs updated with R163 banner
- crates/apeireth-tool-fetch/README.md created (was missing)
- top-level README.md banner added

+++ crates/apeireth-supervisor/src/heartbeat.rs
@@ -351,7 +351,7 @@ impl HeartbeatScheduler {
                                 let mut h = DefaultHasher::new();
                                 item.id.hash(&mut h);
                                 let n = h.finish();
-                                let j = (n % item.schedule.jitter.as_millis() as u64) as u64;
+                                let j = n % item.schedule.jitter.as_millis() as u64;
                                 Duration::from_millis(j)
                             };
                             item.next_tick = Instant::now() + item.schedule.interval + jitter;
